Trying to initiate SNS topic
