+++ ContactDetailsApi.Tests/AwsIntegrationTests.cs
@@ -31,8 +31,6 @@ namespace ContactDetailsApi.Tests
 
         public AwsIntegrationTests()
         {
-            CreateSnsTopic();
-
             EnsureEnvVarConfigured("DynamoDb_LocalMode", "true");
             EnsureEnvVarConfigured("DynamoDb_LocalServiceUrl", "http://localhost:8000");
             EnsureEnvVarConfigured("Localstack_SnsServiceUrl", "http://localhost:4566");
@@ -63,21 +61,6 @@ namespace ContactDetailsApi.Tests
         {
             if (string.IsNullOrEmpty(Environment.GetEnvironmentVariable(name)))
                 Environment.SetEnvironmentVariable(name, defaultValue);
-        }
-
-        private void CreateSnsTopic()
-        {
-            var snsAttrs = new Dictionary<string, string>();
-            snsAttrs.Add("fifo_topic", "true");
-            snsAttrs.Add("content_based_deduplication", "true");
-
-            var response = SimpleNotificationService.CreateTopicAsync(new CreateTopicRequest
-            {
-                Name = "personcreated",
-                Attributes = snsAttrs
-            }).Result;
-
-            Environment.SetEnvironmentVariable("NEW_PERSON_SNS_ARN", response.TopicArn);
         }
     }
